refactor lib.rs into config/errors/optimized_runtime modules, bump to 0.6.0
Split the 2489-line god file into focused modules (config.rs for training
configuration types, errors.rs for all error types, optimized_runtime.rs
absorbing the impl block and all inference helpers), leaving lib.rs as a
thin hub at 189 lines. No functionality changed; all 110 tests pass.
