+++ bindings/python/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "forestfire-python"
-version = "0.5.3"
+version = "0.6.0"
 edition = "2024"
 publish = false
 license = "MIT"
@@ -17,5 +17,5 @@ numpy = "0.26"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 
-forestfire-core = { version = "0.5.3", path = "../../crates/core" }
-forestfire-data = { version = "0.5.3", path = "../../crates/data" }
+forestfire-core = { version = "0.6.0", path = "../../crates/core" }
+forestfire-data = { version = "0.6.0", path = "../../crates/data" }
