+++ crates/inference/Cargo.toml
@@ -7,6 +7,6 @@ repository = "https://github.com/wsperat/forest-fire"
 description = "Inference-facing traits and helpers for ForestFire models."
 
 [dependencies]
-forestfire-core = { version = "0.5.2", path = "../core" }
-forestfire-data = { version = "0.5.2", path = "../data" }
+forestfire-core = { version = "0.6.0", path = "../core" }
+forestfire-data = { version = "0.6.0", path = "../data" }
 arrow.workspace = true
